requirements.txt
torch==2.3.0+cu118
torchvision==0.14.1
pytorch_lightning==1.5.4
beartype==0.18.5
decord==0.6.0
einops==0.8.1
imageio==2.34.2
matplotlib==3.10.1
numpy==2.2.4
opencv_python
pandas==2.2.3
Pillow==11.1.0
Requests==2.32.3
scikit_video==1.1.11
scipy==1.15.2
timm==1.0.15
tqdm
